Refactor player to manage the preview
