+++ projects/packages/videopress/src/client/block-editor/blocks/video/components/player/index.native.js
@@ -2,6 +2,7 @@
  * WordPress dependencies
  */
 import { SandBox } from '@wordpress/components';
+import { useState, useEffect } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 /**
  * External dependencies
@@ -10,18 +11,61 @@ import { View, Text } from 'react-native';
 /**
  * Internal dependencies
  */
+import getMediaToken from '../../../../../lib/get-media-token/index.native';
+import { getVideoPressUrl } from '../../../../../lib/url';
+import { usePreview } from '../../../../hooks/use-preview';
+import addTokenIntoIframeSource from '../../../../utils/add-token-iframe-source';
 import style from './style.scss';
 
 /**
  * VideoPlayer react component
  *
  * @param {object} props - Component props.
- * @param {string} props.html - HTML markup for the player.
- * @param {boolean} props.isRequestingEmbedPreview - Whether the preview is being requested.
+ * @param {object} props.attributes - Block attributes.
  * @param {boolean} props.isSelected - Whether the block is selected.
  * @returns {import('react').ReactElement} - React component.
  */
-export default function Player( { html, isRequestingEmbedPreview, isSelected } ) {
+export default function Player( { isSelected, attributes } ) {
+	const {
+		controls,
+		guid,
+		loop,
+		muted,
+		playsinline,
+		poster,
+		preload,
+		seekbarColor,
+		seekbarLoadingColor,
+		seekbarPlayedColor,
+	} = attributes;
+
+	const [ token, setToken ] = useState();
+
+	// Fetch token for a VideoPress GUID
+	useEffect( () => {
+		if ( guid ) {
+			getMediaToken( 'playback', { guid } ).then( tokenData => {
+				setToken( tokenData.token );
+			} );
+		}
+	}, [ guid ] );
+
+	const videoPressUrl = getVideoPressUrl( guid, {
+		autoplay: false, // Note: Autoplay is disabled to prevent the video from playing fullscreen when loading the editor.
+		controls,
+		loop,
+		muted,
+		playsinline,
+		preload,
+		seekbarColor,
+		seekbarLoadingColor,
+		seekbarPlayedColor,
+		poster,
+	} );
+
+	const { preview, isRequestingEmbedPreview } = usePreview( videoPressUrl );
+	const html = addTokenIntoIframeSource( preview?.html, token );
+
 	// Set up style for when the player is loading.
 	const loadingStyle = {};
 	if ( ! html || isRequestingEmbedPreview ) {
