+++ projects/packages/videopress/src/client/block-editor/blocks/video/edit.native.js
@@ -20,10 +20,7 @@ import { View } from 'react-native';
 /**
  * Internal dependencies
  */
-import getMediaToken from '../../../lib/get-media-token/index.native';
-import { buildVideoPressURL, getVideoPressUrl } from '../../../lib/url';
-import { usePreview } from '../../hooks/use-preview';
-import addTokenIntoIframeSource from '../../utils/add-token-iframe-source';
+import { buildVideoPressURL } from '../../../lib/url';
 import isLocalFile from '../../utils/is-local-file';
 import ColorPanel from './components/color-panel';
 import DetailsPanel from './components/details-panel';
@@ -54,18 +51,7 @@ export default function VideoPressEdit( {
 	onFocus,
 	insertBlocksAfter,
 } ) {
-	const {
-		controls,
-		guid,
-		loop,
-		muted,
-		playsinline,
-		poster,
-		preload,
-		seekbarColor,
-		seekbarLoadingColor,
-		seekbarPlayedColor,
-	} = attributes;
+	const { guid } = attributes;
 
 	const [ isUploadingFile, setIsUploadingFile ] = useState( ! guid );
 	const [ fileToUpload, setFileToUpload ] = useState( null );
@@ -73,16 +59,6 @@ export default function VideoPressEdit( {
 		isReplacing: false,
 		prevAttrs: {},
 	} );
-	const [ token, setToken ] = useState();
-
-	// Fetch token for a VideoPress GUID
-	useEffect( () => {
-		if ( guid ) {
-			getMediaToken( 'playback', { guid } ).then( tokenData => {
-				setToken( tokenData.token );
-			} );
-		}
-	}, [ guid ] );
 
 	const [ showReplaceControl, setShowReplaceControl ] = useState( true );
 
@@ -92,22 +68,6 @@ export default function VideoPressEdit( {
 	);
 	const { replaceBlock } = useDispatch( blockEditorStore );
 	const { createErrorNotice } = useDispatch( noticesStore );
-
-	const videoPressUrl = getVideoPressUrl( guid, {
-		autoplay: false, // Note: Autoplay is disabled to prevent the video from playing fullscreen when loading the editor.
-		controls,
-		loop,
-		muted,
-		playsinline,
-		preload,
-		seekbarColor,
-		seekbarLoadingColor,
-		seekbarPlayedColor,
-		poster,
-	} );
-
-	const { preview, isRequestingEmbedPreview } = usePreview( videoPressUrl );
-	const previewHTML = addTokenIntoIframeSource( preview?.html, token );
 
 	// Display upload progress in case the editor is closed and re-opened
 	// while the upload is in progress.
@@ -255,11 +215,7 @@ export default function VideoPressEdit( {
 				</InspectorControls>
 			) }
 
-			<Player
-				html={ previewHTML }
-				isRequestingEmbedPreview={ isRequestingEmbedPreview }
-				isSelected={ isSelected }
-			/>
+			<Player { ...{ attributes, isSelected } } />
 
 			<BlockCaption
 				clientId={ clientId }
